+++ actionpack/test/dispatch/host_authorization_test.rb
@@ -259,12 +259,6 @@ class HostAuthorizationTest < ActionDispatch::IntegrationTest
     assert_match "Blocked host: sub-example.com", response.body
   end
 
-  test "config setting action_dispatch.hosts_response_app is deprecated" do
-    assert_deprecated do
-      ActionDispatch::HostAuthorization.new(App, "example.com", ->(env) { true })
-    end
-  end
-
   test "uses logger from the env" do
     @app = ActionDispatch::HostAuthorization.new(App, %w(only.com))
     output = StringIO.new